requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-Limiter==3.12
prometheus-flask-exporter==0.23.2
python-dotenv==1.0.1
markdown==3.7
requests==2.32.3
structlog==24.4.0
gunicorn==23.0.0
